% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desplot.R
\name{RedGrayBlue}
\alias{RedGrayBlue}
\title{Function to create a Red-Gray-Blue palette}
\usage{
RedGrayBlue(n)
}
\arguments{
\item{n}{Number of colors to create}
}
\value{
A vector of n colors.
}
\description{
A function to create a Red-Gray-Blue palette.
}
\details{
Using gray instead of white allows missing values to appear as white
(actually, transparent).
}
\examples{
pie(rep(1,11), col=RedGrayBlue(11))
title("RedGrayBlue(11)")
}
\author{
Kevin Wright
}
